+++ mitra_models/migrations/schema.sql
@@ -367,6 +367,7 @@ CREATE TABLE invoice (
     payment_type SMALLINT,
     payment_address VARCHAR(500),
     payout_tx_id VARCHAR(200),
+    payout_amount BIGINT CHECK (payout_amount > 0),
     object_id VARCHAR(2000) UNIQUE,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -390,6 +391,7 @@ CREATE TABLE notification (
     recipient_id UUID NOT NULL REFERENCES user_account (id) ON DELETE CASCADE,
     post_id UUID REFERENCES post (id) ON DELETE CASCADE,
     reaction_id UUID REFERENCES post_reaction (id) ON DELETE CASCADE,
+    invoice_id UUID REFERENCES invoice (id) ON DELETE CASCADE,
     event_type SMALLINT NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now()
 );